Getting places

+++ packages/entity-composer/src/react.tsx
@@ -8,11 +8,13 @@ import React, {
 } from "react"
 import { Bucket, IEntity } from "./bucket"
 import { useConst } from "@hmans/use-const"
+import { useRerender } from "@hmans/use-rerender"
 
 const useIsomorphicLayoutEffect =
   typeof window !== "undefined" ? useLayoutEffect : useEffect
 
 export type EntityProps<E extends IEntity> = Partial<E> & {
+  entity?: E
   children?: ReactNode
 }
 
@@ -25,13 +27,18 @@ export type PropertyProps<E extends IEntity, P extends keyof E> = {
 export const createComponents = <E extends IEntity>(bucket: Bucket<E>) => {
   const EntityContext = createContext<E | null>(null)
 
-  const Entity = ({ children, ...props }: EntityProps<E>) => {
-    const entity = useConst(() => ({} as E))
+  const Entity = ({
+    children,
+    entity: existingEntity,
+    ...props
+  }: EntityProps<E>) => {
+    const entity = useConst(() => existingEntity ?? ({} as E))
 
     useIsomorphicLayoutEffect(() => {
       bucket.write(entity, props as Partial<E>)
+      if (existingEntity) return
       return () => bucket.remove(entity)
-    })
+    }, [])
 
     return (
       <EntityContext.Provider value={entity}>{children}</EntityContext.Provider>
@@ -40,6 +47,10 @@ export const createComponents = <E extends IEntity>(bucket: Bucket<E>) => {
 
   const Property = <P extends keyof E>(props: PropertyProps<E, P>) => {
     const entity = useContext(EntityContext)
+
+    if (!entity) {
+      throw new Error("Property must be a child of Entity")
+    }
 
     /* Handle setting of value */
     useIsomorphicLayoutEffect(() => {
@@ -54,12 +65,32 @@ export const createComponents = <E extends IEntity>(bucket: Bucket<E>) => {
     const children = props.children
       ? React.cloneElement(
           React.Children.only(props.children) as ReactElement,
-          { ref: (ref: E[P]) => (entity![props.name] = ref) }
+          {
+            ref: (ref: E[P]) => {
+              bucket.write(entity, { [props.name]: ref })
+            }
+          }
         )
       : null
 
     return <>{children}</>
   }
 
-  return { Entity, Property }
+  return { Entity, Property, useBucket }
+}
+
+export const useBucket = <E extends IEntity>(bucket: Bucket<E>) => {
+  const rerender = useRerender()
+
+  useIsomorphicLayoutEffect(() => {
+    bucket.onEntityAdded.addListener(rerender)
+    bucket.onEntityRemoved.addListener(rerender)
+
+    return () => {
+      bucket.onEntityAdded.removeListener(rerender)
+      bucket.onEntityRemoved.removeListener(rerender)
+    }
+  }, [])
+
+  return bucket
 }
